#1 fix: entity type & naming 변경

+++ src/main/java/smile/iceBulterrecipe/recipe/entity/Cookery.java
@@ -15,7 +15,7 @@ public class Cookery extends BaseEntity {
     @Column(nullable = false)
     private Long cookeryIdx;
 
-    private int nextIdx;
+    private Long nextIdx;
     private String cookeryImg;
     private String description;
